Fixes permissions with all Android versions Fixes how we are granting permissions in the app. Adds onTaskRemoved in WifiListener to start it again if Android closes it. Deprecates Android Oreo and below. Removes deprecated code. Updates libraries to latest. Adds readme file
